[move/#134] confirm deny type 패키지 이동
